Cast the narrowings, turn the flags on, and make the disk check assert the interlock
The six C4267 warnings hbios_dispatch.cc emits under MSVC at /W3 are a
size_t loop index promoting a uint16_t guest address, where truncation to
sixteen bits IS the Z80 64K wrap HBIOS wants. Each site gained an explicit
cast that says so, and -Wimplicit-int-conversion is on - guarded by a
$(shell) probe the way cpmemu guards -Wshorten-64-to-32, because GCC does
not ignore an unknown -W flag, it fails the compile, and a hardcoded flag
has stopped every ubuntu job in this family before. romwbw_emu.cc's two
-Wshorten-64-to-32 sites went the same way and that flag is on too. Neither
was a live bug; they were the only thing standing between this tree and a
warning-free build under either compiler, which is what makes /WX possible
in the msvc job later.

disks/verify_disk_utils.sh compared the shipped r8.com and w8.com against
their sources but could not tell an armed W8 from an interlocked one - the
"Usage: W8 <cpmname> [hostpath]" string is identical in both, and three
armed blobs exist in this repo's history. It now asserts the HBF_HOST_CAPS
probe bytes 06 e9 cf, which is what RELEASE_ORDER asked for and what stops a
stale image being attached by mistake. Verified to bite: an armed w8.com
extracted from de85946 fails the new check.

roms/verify_romwbw_pin.sh pruned only .git and archive, and both CI jobs
clone cpmemu into the workspace root, so it walked cpmemu/tests/*.bin and
reported 17 mismatches against RomWBW v3.5.1 on Z80 blobs that were never
meant to carry an HCB. Anyone following the documented CI recipe locally hit
it; it also blocked ever adding the script to the workflow.

tests/cli_console.cc is back in the macos job. It was dropped for a
flakiness the file's own comments in the same tree describe as fixed - the
race was an assertion running after the parent's write, not the pty round
trip - and it measured 0 failures in 390 runs here.

docs/RELEASE_ORDER_2026-08-25.md was substantially stale and three other
repos point at it. Step 0's "committed locally, NOT pushed" is false (v1.36
is pushed and main is five commits past it, and both tracked images carry
the interlocked W8); step 1's ioscpm fix is committed; step 2's cpmdroid is
checked out and done. It gained a section for the second data-loss path on
step 5, which it did not mention at all: ioscpm's catalog-version check
deletes every .img in Documents/Disks, including ones the user imported and
ones the app made, and step 4 is what fires it.

todo.txt is 269 lines to 127, tagged by the machine that can take each item.

+++ src/emu_init.cc
@@ -258,7 +258,7 @@ void emu_copy_hcb_to_shadow_ram(banked_mem* memory) {
   memory->select_bank(0x00);  // ROM bank 0 mode
 
   for (int i = 0; i < 512; i++) {
-    memory->store_mem(i, rom[i]);  // Writes to shadow RAM and sets shadow bit
+    memory->store_mem((uint16_t)i, rom[i]);  // Writes to shadow RAM and sets shadow bit
   }
 
   memory->select_bank(saved_bank);  // Restore previous bank
@@ -383,7 +383,7 @@ int emu_populate_drive_map(banked_mem* memory, HBIOSDispatch* hbios,
   // First, mark all drive map entries as unused (0xFF) in both ROM and RAM
   for (int i = 0; i < 16; i++) {
     rom[DRVMAP_BASE + i] = 0xFF;
-    memory->write_bank(0x80, DRVMAP_BASE + i, 0xFF);
+    memory->write_bank(0x80, (uint16_t)(DRVMAP_BASE + i), 0xFF);
   }
 
   // IMPORTANT: Hard disks are assigned FIRST so boot disk is A:
@@ -403,9 +403,9 @@ int emu_populate_drive_map(banked_mem* memory, HBIOSDispatch* hbios,
 
         // Assign each slice to a drive letter
         for (int slice = 0; slice < num_slices && drive_letter < 16; slice++) {
-          uint8_t map_value = ((slice & 0x0F) << 4) | (unit & 0x0F);
+          uint8_t map_value = (uint8_t)(((slice & 0x0F) << 4) | (unit & 0x0F));
           rom[DRVMAP_BASE + drive_letter] = map_value;
-          memory->write_bank(0x80, DRVMAP_BASE + drive_letter, map_value);
+          memory->write_bank(0x80, (uint16_t)(DRVMAP_BASE + drive_letter), map_value);
           emu_log("[EMU_INIT] Drive %c: = HDSK%d:%d (unit=%d, map=0x%02X)\n",
                   'A' + drive_letter, hd, slice, unit, map_value);
           drive_letter++;
@@ -418,7 +418,7 @@ int emu_populate_drive_map(banked_mem* memory, HBIOSDispatch* hbios,
   // MD0 (RAM disk) if enabled
   if (ramd_banks > 0 && drive_letter < 16) {
     rom[DRVMAP_BASE + drive_letter] = 0x00;  // Unit 0, slice 0
-    memory->write_bank(0x80, DRVMAP_BASE + drive_letter, 0x00);
+    memory->write_bank(0x80, (uint16_t)(DRVMAP_BASE + drive_letter), 0x00);
     emu_log("[EMU_INIT] Drive %c: = MD0 (RAM disk)\n", 'A' + drive_letter);
     drive_letter++;
   }
@@ -426,7 +426,7 @@ int emu_populate_drive_map(banked_mem* memory, HBIOSDispatch* hbios,
   // MD1 (ROM disk) if enabled
   if (romd_banks > 0 && drive_letter < 16) {
     rom[DRVMAP_BASE + drive_letter] = 0x01;  // Unit 1, slice 0
-    memory->write_bank(0x80, DRVMAP_BASE + drive_letter, 0x01);
+    memory->write_bank(0x80, (uint16_t)(DRVMAP_BASE + drive_letter), 0x01);
     emu_log("[EMU_INIT] Drive %c: = MD1 (ROM disk)\n", 'A' + drive_letter);
     drive_letter++;
   }
@@ -596,8 +596,8 @@ void emu_copy_rom_apps_to_ram(banked_mem* memory) {
           app_count, ROM_APP_START_BANK, app_ram_start);
 
   for (int i = 0; i < app_count; i++) {
-    uint8_t rom_bank = ROM_APP_START_BANK + i;
-    uint8_t ram_bank = app_ram_start + i;
+    uint8_t rom_bank = (uint8_t)(ROM_APP_START_BANK + i);
+    uint8_t ram_bank = (uint8_t)(app_ram_start + i);
 
     // Copy entire 32KB bank from ROM to RAM
     for (uint16_t addr = 0; addr < 0x8000; addr++) {
